--- a/src/Delimiter/Processor/StaggeredDelimiterProcessor.php
+++ b/src/Delimiter/Processor/StaggeredDelimiterProcessor.php
@@ -74,13 +74,13 @@ final class StaggeredDelimiterProcessor implements DelimiterProcessorInterface
         $len = $processor->getMinLength();
 
         if (isset($this->processors[$len])) {
-            throw new \InvalidArgumentException(sprintf('Cannot add two delimiter processors for char "%s" and minimum length %d', $this->delimiterChar, $len));
+            throw new \InvalidArgumentException(\sprintf('Cannot add two delimiter processors for char "%s" and minimum length %d', $this->delimiterChar, $len));
         }
 
         $this->processors[$len] = $processor;
-        krsort($this->processors);
+        \krsort($this->processors);
 
-        $this->minLength = min($this->minLength, $len);
+        $this->minLength = \min($this->minLength, $len);
     }
 
     /**
@@ -110,7 +110,7 @@ final class StaggeredDelimiterProcessor implements DelimiterProcessorInterface
 
         // Just use the first one in our list
         /** @var DelimiterProcessorInterface $first */
-        $first = reset($this->processors);
+        $first = \reset($this->processors);
 
         return $first;
     }