--- a/src/Inline/Parser/CloseBracketParser.php
+++ b/src/Inline/Parser/CloseBracketParser.php
@@ -194,10 +194,17 @@ final class CloseBracketParser implements InlineParserInterface, EnvironmentAwar
         $beforeLabel = $cursor->getPosition();
         $n = LinkParserHelper::parseLinkLabel($cursor);
         if ($n === 0 || $n === 2) {
-            // Empty or missing second label
-            $reflabel = \mb_substr($cursor->getLine(), $opener->getIndex(), $startPos - $opener->getIndex(), 'utf-8');
+            $start = $opener->getIndex();
+            $length = $startPos - $opener->getIndex();
         } else {
-            $reflabel = \mb_substr($cursor->getLine(), $beforeLabel + 1, $n - 2, 'utf-8');
+            $start = $beforeLabel + 1;
+            $length = $n - 2;
+        }
+
+        if ($cursor->isMultiByte()) {
+            $referenceLabel = \mb_substr($cursor->getLine(), $start, $length, $cursor->getEncoding());
+        } else {
+            $referenceLabel = \substr($cursor->getLine(), $start, $length);
         }
 
         if ($n === 0) {
@@ -205,7 +212,7 @@ final class CloseBracketParser implements InlineParserInterface, EnvironmentAwar
             $cursor->restoreState($savePos);
         }
 
-        return $referenceMap->getReference($reflabel);
+        return $referenceMap->getReference($referenceLabel);
     }
 
     /**
@@ -215,7 +222,7 @@ final class CloseBracketParser implements InlineParserInterface, EnvironmentAwar
      *
      * @return AbstractWebResource
      */
-    private function createInline(string $url, string $title, bool $isImage)
+    private function createInline(string $url, string $title, bool $isImage): AbstractWebResource
     {
         if ($isImage) {
             return new Image($url, null, $title);